Code fixes and memory check seems to be working

+++ src/main/scala/autoscale/model/MarathonTask.scala
@@ -22,7 +22,7 @@ case class Statistics(cpusSystemTimeSecs: Float,
                       cpusUserTimeSecs:Float,
                       memLimitBytes:Float,
                       memRssBytes:Float,
-                      timestamp: Float)
+                      timestamp: Double)
 
 object Statistics{
   val taskStatisticsRename = FieldSerializer[Statistics](
